refactor: extract the applied-check read path from the bridge module
extension_bridge/mod.rs grew past the R8 hard LOC cap (tests/architecture.rs)
once the live-connection refcount landed. Move the self-contained
applied.check machinery (AppliedCheckOk, resolve_applied_check,
applied_result_reply, handle_applied_check) into a sibling applied_check.rs,
mirroring the existing status_update.rs / autotrack.rs split. Pure move: no
behavior change, pub(super) visibility so import_tests.rs keeps direct
access, call sites updated to applied_check::handle_applied_check.

mod.rs: 1419 -> 1316 LOC (cap is 1400).
